// src/app/modules/lshp-admin/models/parent-data.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ParentData
 */
export interface ParentData { 
    /**
     * Gets or Sets otp
     */
    otp?: number;
    /**
     * Gets or Sets ParentId
     */
    parentId?: number;
    /**
     * Gets or Sets Email
     */
    email?: string;
    /**
     * Gets or Sets SchoolId
     */
    schoolId?: string;
     /**
     * Gets or Sets mobileNumber
     */
    mobileNumber?: string;
     /**
     * Gets or Sets AdmissionNumber
     */
    admissionNumber?: string;
     /**
     * Gets or Sets Password
     */
    password?: string;

    /**
    * Gets or Sets Password
    */
    schoolCode?: string;
}