// src/app/modules/communication/model/mobile-number-view.ts
/**
 * Communication
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * MobileNumberView
 */
export interface MobileNumberView { 
    /**
     * Gets or Sets MobileNumber
     */
    mobileNumber?: string;

    /**
     * Gets or Sets schoolId
     */
    schoolId?: string;
}