// src/app/modules/lshp-admin/models/emailResendModel.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * EmailResendModel
 */
export interface EmailResendModel { 
    /**
     * Gets or Sets Email
     */
    email?: string;
    /**
     * Gets or Sets SchoolId
     */
    schoolId?: string;
}